- compile warning fix
svn path=/trunk/boinc/; revision=24461

+++ lib/coproc.cpp
@@ -262,7 +262,7 @@ void OPENCL_DEVICE_PROP::description(char* buf, const char* type) {
     
     // openCL_device_version may have a trailing space
     strlcpy(s1, opencl_device_version, sizeof(s1));
-    n = strlen(s1) - 1;
+    n = (int)strlen(s1) - 1;
     if ((n > 0) && (s1[n] == ' ')) s1[n] = '\0';
     sprintf(s2, "%s (driver version %s, device version %s, %.0fMB)",
         name, opencl_driver_version, s1, global_mem_size/MEGA
